Don't use AI_ADDRCONFIG for bind addr translation
Fixes #52

On Linux, the getaddrinfo function returns EAI_NONAME when node is NULL,
the AI_ADDRCONFIG flag is set, and the system only has loopback
addresses configured.

Quoting getaddrinfo(3):
> If hints.ai_flags includes the AI_ADDRCONFIG flag, then IPv4
> addresses are returned in the list pointed to by res only if the
> local system has at least one IPv4 address configured, and IPv6
> addresses are returned only if the local system has at least one
> IPv6 address configured.  The loopback address is not considered
> for this case as valid as a configured address.  This flag is
> useful on, for example, IPv4-only systems, to ensure that
> getaddrinfo() does not return IPv6 socket addresses that would
> always fail in connect(2) or bind(2).

Note that the FreeBSD version of getaddrinfo(3) does not mention that
the loopback address is not considered a valid configured address.

It is desirable that binding to wildcard addresses (the default) also
works when no routable IP addresses are configured on the system. This
can be achieved by removing the AI_ADDRCONFIG flag.

The flag appears to have been set in order to optimize DNS query
traffic [1]. My testing however shows that no DNS querying takes place
as a result of anything done in the wait_for_connections function, so I
believe that this change has no performance impact.

[1] https://fedoraproject.org/wiki/QA/Networking/NameResolution/ADDRCONFIG